requirements.txt
pandas
numpy
PyYAML
pathlib
ipykernel
notebook
python-box
ensure
spacy
neo4j
sentence-transformers
faiss-cpu
google-generativeai
pypdf

# LangChain RAG Stack
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
tiktoken
langchain-groq
# Additional dependencies for enhanced RAG
python-dotenv


chainlit

#evaluation
deepeval
datasets